+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Fix Fabric dependency showing for mods even when installed on another mod platform
 - Fix issue with Quilt mods showing as updates for Fabric instances [#622]
 - Don't show Fabric dependency when QSL is installed
+- Issue when checking for updates on Modrinth not using correct loaders
 
 ### Misc
 - Switch to using Log4j2 for application logging [#533] (thanks @s0cks)